penamabahan modal delete

+++ resources/views/CreateProfile.blade.php
@@ -189,7 +189,8 @@
                                                       <td>
                                                           <a href="#" type="button" class="btn btn-outline-primary btn-sm" onclick="viewDetail('{{ $datas['ID'] }}')">Detail</a>
                                                           <a href="{{ route('profile.history', ['id' =>$datas['ID']]) }}" type="button" class="btn btn-outline-info btn-sm" >history</a>
-                                                          <a href="{{ route('profile.drop', ['id' =>$datas['ID']]) }}" type="delete" class="btn btn-outline-danger btn-sm" >Delete</a>
+                                                          <a href="{{ route('profile.drop', ['id' =>$datas['ID']]) }}" type="delete" class="btn btn-outline-danger btn-sm" data-href="/delete.php?id=54" data-toggle="modal" data-target="#confirm-delete" >Delete</a>
+                                                          <!-- <button class="btn btn-default" data-href="/delete.php?id=54" data-toggle="modal" data-target="#confirm-delete"> -->
                                                       </td>
                                                   </tr>
                                               @endforeach
@@ -301,7 +302,7 @@
                                                                 <div class="form-group row">
                                                                     <p class="col-sm-3 col-form-label">ZipCode</p>
                                                                         <div class="col-sm-3">
-                                                                            <input class="form-control" id="ZipCode" name="ZipCode" type="text">
+                                                                            <input class="form-control" id="ZipCode" name="ZipCode" type="text" maxlength="10">
                                                                         </div>
                                                                 </div>
                                                                 <div class="form-group row">
@@ -568,7 +569,7 @@
                                                           <div class="form-group row">
                                                               <p class="col-sm-3 col-form-label">ZipCode</p>
                                                               <div class="col-sm-3">
-                                                                  <input class="form-control" id="TxtProfileZipCode" name="ZipCode" type="text" value="{{ old('ZipCode') }}">
+                                                                  <input class="form-control" id="TxtProfileZipCode" name="ZipCode" type="text" value="{{ old('ZipCode') }}" maxlength="10">
                                                               </div>
                                                           </div>
                                                           <div class="form-group row">
@@ -751,6 +752,29 @@
                   </div>
               </div>
           </div>
+          <!-- cobain modal delete -->
+          <div class="modal fade" id="confirm-delete" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+        <div class="modal-dialog">
+            <div class="modal-content">
+            
+                <div class="modal-header">
+                    <h4 class="modal-title" id="myModalLabel">Confirm Delete</h4>
+                </div>
+            
+                <div class="modal-body">
+                    <p>You are about to delete one track, this procedure is irreversible.</p>
+                    <p>Do you want to proceed?</p>
+                    <p class="debug-url"></p>
+                </div>
+                
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
+                    <a class="btn btn-danger btn-ok">Delete</a>
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- batas bawah -->
         </main>
     </section>
       
@@ -807,6 +831,15 @@
 <script src="{{asset('plugins/toastr/toastr.min.js')}}"></script>
 <!-- SweetAlert2 -->
 <script src="{{asset('plugins/sweetalert2/sweetalert2.min.js')}}"></script>
+
+<!-- script modal button delete -->
+<script>
+        $('#confirm-delete').on('show.bs.modal', function(e) {
+            $(this).find('.btn-ok').attr('href', $(e.relatedTarget).data('href'));
+            
+            $('.debug-url').html('Delete URL: <strong>' + $(this).find('.btn-ok').att    r('href') + '</strong>');
+        });
+    </script>
 
 <script>
 function CGroup_OnChange(CGroup){
@@ -1049,10 +1082,10 @@ function GetFormattedDate(datestring) {
 </script>
 <script>
 $('#search').click(function(event) {
-    $("#tblModalSync").empty();//Clear old data before ajax
+    $("#tblModalSync").empty();//hapus data lama 
     $.ajax({
         url : 'http://uat2.care.co.id:9095/aca/WEBAPI2/MiddlewareAPI/SearchHistoryProfile',
-        type : 'GET',
+        type : 'POST',
         dataType : 'json',
         success : function(data) {
             var table = $("#tblModalSync");
